+++ local-store/ClearFullStore.js
@@ -5,31 +5,37 @@ var UniqueChronologicalKeys = require("./UniqueChronologicalKeys");
  * @since 2016-3-25
  */
 var ClearFullStore = (function () {
-    function ClearFullStore(extractChronoId, removePercentage) {
+    function ClearFullStore(extractChronoId, itemsRemovedCallback, removePercentage) {
         if (removePercentage === void 0) { removePercentage = 0.3; }
         /** [0.0, 1.0] when the local store gets too full, remove this percentage of the entries */
         this.removePercentage = 0.3;
         /** an internal counter of how many times clearOldItems() has been called */
         this.removalAttemptsCount = 0;
         this.extractChronoId = extractChronoId;
+        this.itemsRemovedCallback = itemsRemovedCallback;
         this.removePercentage = removePercentage;
     }
     /** Remove old items
      */
-    ClearFullStore.prototype.clearOldItems = function (storeInst, logInfo, err, removePercentage, minItemsRemoved, maxItemsRemoved) {
+    ClearFullStore.prototype.clearOldItems = function (store, logInfo, err, removePercentage, minItemsRemoved, maxItemsRemoved) {
         if (removePercentage === void 0) { removePercentage = this.removePercentage; }
         if (minItemsRemoved === void 0) { minItemsRemoved = 1; }
         if (logInfo) {
-            var start = window.performance.now();
+            var start = UniqueChronologicalKeys.getMillisecondTime();
         }
-        var removeCount = ClearFullStore.removeOldItems(storeInst, this.extractChronoId, removePercentage, minItemsRemoved, maxItemsRemoved);
+        var removedItems = ClearFullStore.removeOldItems(store, this.extractChronoId, removePercentage, minItemsRemoved, maxItemsRemoved);
         this.removalAttemptsCount++;
         if (logInfo) {
-            var end = window.performance.now();
-            // TODO poor solution, log this manually, since psLog imports this class and we don't want a circular dependency
-            storeInst.setItem(UniqueChronologicalKeys.uniqueTimestamp() + "", "removed " + removeCount + " local store entries in " + Math.round(end - start) + " ms, because local store threw error" + (err ? ": '" + err.message + "': " + JSON.stringify(err.stack) : ""));
+            var end = UniqueChronologicalKeys.getMillisecondTime();
         }
-        return removeCount;
+        if (this.itemsRemovedCallback) {
+            this.itemsRemovedCallback(store, removedItems.items, {
+                error: err,
+                message: "removed " + removedItems.items.length + " local store entries" + (start || end ? " in " + Math.round(end - start) + " ms" : "") + ", " +
+                    "because local store threw error" + (err ? ": '" + err.message + "': " + JSON.stringify(err.stack) : ""),
+            }, removedItems.removedCount);
+        }
+        return removedItems.items;
     };
     /** Given a {@link LocalStore} with integer based keys (not all keys must be integers),
      * remove {@code removePercentage} percent of the smallest integer keys
@@ -42,28 +48,39 @@ var ClearFullStore = (function () {
      */
     ClearFullStore.removeOldItems = function (localStore, extractChronoId, removePercentage, minItemsRemoved, maxItemsRemoved) {
         if (minItemsRemoved === void 0) { minItemsRemoved = 1; }
-        var ids = [];
-        // get a list of integer keys from the local store
+        var idsKeys = [];
+        // get a list of chronological keys from the local store
         var keys = localStore.getKeys();
         for (var i = 0, size = keys.length; i < size; i++) {
             var val = extractChronoId(keys[i]);
             if (Number.isInteger(val)) {
-                ids.push(val);
+                idsKeys.push({
+                    id: val,
+                    key: keys[i]
+                });
             }
         }
         // sort ascending
-        ids.sort(function (a, b) { return a - b; });
-        var idCount = ids.length;
+        idsKeys.sort(function (a, b) { return a.id - b.id; });
+        var idCount = idsKeys.length;
         maxItemsRemoved = maxItemsRemoved === undefined ? idCount : Math.min(idCount, maxItemsRemoved);
         var removeCount = Math.max(Math.min(Math.round(idCount * removePercentage), maxItemsRemoved), minItemsRemoved);
+        var removedItems = [];
         // remove the oldest timestamped entries (always remove between [1, timestamps.length] entries)
         for (var i = 0; i < removeCount; i++) {
-            localStore.removeItem(ids[i].toString());
+            var keyId = idsKeys[i];
+            var existingItem = localStore.getItem(keyId.key);
+            removedItems.push({
+                key: idsKeys[i].key,
+                keyId: idsKeys[i].id,
+                value: existingItem,
+            });
+            localStore.removeItem(keyId.key);
         }
-        return removeCount;
+        return { items: removedItems, idsSorted: idsKeys, removedCount: removeCount };
     };
-    ClearFullStore.newInst = function (extractChronoId, removePercentage) {
-        return new ClearFullStore(extractChronoId, removePercentage);
+    ClearFullStore.newInst = function (extractChronoId, itemsRemovedCallback, removePercentage) {
+        return new ClearFullStore(extractChronoId, itemsRemovedCallback, removePercentage);
     };
     return ClearFullStore;
 }());