+++ test/local-store/LocalStoreByCategoryTest.ts
@@ -15,7 +15,7 @@ interface CategoryEvent {
 }
 
 var memStore = MemoryStore.newInst();
-var localStore = LocalStoreFromStorage.newInst(memStore, null, true, false, 80, false);
+var localStore = LocalStoreFromStorage.newInst(memStore, null, null, true, false, 80, false);
 var storeBldr = LocalStoreByCategory.Builder.newInst(localStore, UniqueChronologicalKeys.uniqueTimestampNodeJs);
 var store = storeBldr.addStores({
     alpha: storeBldr.toStore(BasicCategorizers.newPrefixCategorizer("alpha-")),